# Rewrite Kernel#loop in Ruby