because Ruby is faster than C ™️

## Interpreter

```
$ benchmark-driver -v --chruby 'before;after' benchmark/loop_generator.rb
before: ruby 3.2.0dev (2022-12-21T15:34:21Z master 398aaed2f0) [arm64-darwin22]
after: ruby 3.2.0dev (2022-12-21T18:57:36Z ruby-loop b860fd41db) [arm64-darwin22]
Calculating -------------------------------------
                         before       after
      loop_generator      0.888       0.943 i/s -       1.000 times in 1.125595s 1.060871s

Comparison:
                   loop_generator
               after:         0.9 i/s
              before:         0.9 i/s - 1.06x  slower
```

## YJIT

```
benchmark-driver -v --chruby 'before::before --yjit-call-threshold=1;after::after --yjit-call-threshold=1' benchmark/loop_generator.rb
before: ruby 3.2.0dev (2022-12-21T15:34:21Z master 398aaed2f0) +YJIT [arm64-darwin22]
after: ruby 3.2.0dev (2022-12-21T18:57:36Z ruby-loop b860fd41db) +YJIT [arm64-darwin22]
Calculating -------------------------------------
                         before       after
      loop_generator      0.962       1.146 i/s -       1.000 times in 1.039119s 0.872700s

Comparison:
                   loop_generator
               after:         1.1 i/s
              before:         1.0 i/s - 1.19x  slower
```